requirements.txt
django==3.0.8
djangorestframework==3.11.0
pillow==7.2.0